Add files via upload
This script aligns multiple images based on feature matching and stacks them into a single multi-dimensional array. The output is saved in both TIFF (for general use) and ENVI (specifically for hyperspectral imaging data) formats. This is particularly designed for multispectral imaging in cultural heritage applications. This script is designed for low-cost portable systems, which may take each individual image as a separate file.